User data and daily calorie consumption on graph (#7)
* Added chart to home view

* Added chart

* Styled chart

* Added saving user data

* added target daily calorie visualisation

* add daily macros

* added circular graph

+++ NutritionApp/Search/SearchPresenter.swift
@@ -30,7 +30,7 @@ class SearchPresenter: ObservableObject {
     }
 
     func update(servingSize: String, for nutritionItem: NutritionalItemViewModel) {
-        let value = CGFloat(Float(servingSize) ?? 0)
+        let value = Float(servingSize) ?? 0
 
         meal = meal.update(servingSize: value, for: nutritionItem)
     }
